1.2

+++ DebtDiary/Model/CoreData-Helper.swift
@@ -10,7 +10,7 @@ import Foundation
 extension Cash: Comparable {
     var amount: Int {
         get { Int(cashAmount) }
-        set { cashAmount = Int16(newValue) }
+        set { cashAmount = Int64(newValue) }
     }
     
     var title: String {
